requirements.txt
alembic==0.9.5
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
bcrypt==3.1.3
cffi==1.10.0
click==6.7
coverage==4.4.1
cryptography==2.7
entrypoints==0.3
flake8==3.7.7
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.0
Flask-Script==2.0.5
Flask-SocketIO==4.1.0
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
ForgeryPy==0.1
gunicorn==19.9.0
isort==4.3.20
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.4.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
nose==1.3.7
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.18
pyflakes==2.1.1
PyJWT==1.5.2
pylint==2.3.1
python-dateutil==2.6.1
python-editor==1.0.3
python-engineio==3.8.1
python-socketio==4.1.0
six==1.10.0
SQLAlchemy==1.1.13
typed-ast==1.3.5
Werkzeug==0.12.2
wrapt==1.11.2
WTForms==2.2.1
